+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -334,7 +334,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWG6JW4VRKA5JCRBNNEFRY01"
+        "01KWGTD7S082G067ZJ8J3KKQ12"
     )
     r"""The unique identifier of the data source"""