+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsefunction.py
@@ -507,9 +507,9 @@ class EvaluatorResponseFunction(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-02T01:19:40.024Z"
+    created: Optional[str] = "2026-07-02T07:06:06.701Z"
 
-    updated: Optional[str] = "2026-07-02T01:19:40.024Z"
+    updated: Optional[str] = "2026-07-02T07:06:06.701Z"
 
     updated_by_id: OptionalNullable[str] = UNSET