+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsetypescript.py
@@ -40,9 +40,9 @@ class EvaluatorResponseTypescript(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-02T01:19:40.024Z"
+    created: Optional[str] = "2026-07-02T07:06:06.701Z"
 
-    updated: Optional[str] = "2026-07-02T01:19:40.024Z"
+    updated: Optional[str] = "2026-07-02T07:06:06.701Z"
 
     updated_by_id: OptionalNullable[str] = UNSET