+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWG6JW4S2WTMVP95B0GG2B3Z"
+        "01KWGTD7RZHHN6H7H2N5DKDX1M"
     )
     r"""The unique identifier of the data source"""