+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWG6JW4ZZ0TPMK3Z5DNM9Z74"
+        "01KWGTD7S4DNTMPFBEJCGJE9HD"
     )
     r"""The unique identifier of the data source"""