+++ packages/orq-rc/RELEASES.md
@@ -9179,3 +9179,13 @@ Based on:
 - [python v4.11.0-rc.68] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.68] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.68 - packages/orq-rc
+
+## 2026-07-02 07:08:47
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.787.0 (2.914.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.69] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.69] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.69 - packages/orq-rc